Added Basic Details to Home Fragment

+++ app/src/main/java/com/tec9ers/thunderstorm/model/onecallapi/Minutely.kt
@@ -8,5 +8,5 @@ data class Minutely (
 	val dt : Int,
 
 	@SerializedName("precipitation")
-	val precipitation : Int
+	val precipitation : Double
 )
